Move HG recipes to its own directory
