+++ src/fairseq2/datasets/__init__.py
@@ -22,3 +22,4 @@ from fairseq2.datasets.loader import is_dataset_card as is_dataset_card
 import fairseq2.datasets.asr
 import fairseq2.datasets.instruction
 import fairseq2.datasets.parallel_text
+import fairseq2.datasets.text
