+++ src/fairseq2/recipes/__init__.py
@@ -16,7 +16,7 @@ import os
 
 from fairseq2.logging import get_log_writer
 from fairseq2.recipes.assets import _setup_asset_cli
-from fairseq2.recipes.eval import _setup_eval_cli
+from fairseq2.recipes.hg import _setup_hg_cli
 from fairseq2.recipes.llama import _setup_llama_cli
 from fairseq2.recipes.lm import _setup_lm_cli
 from fairseq2.recipes.logging import setup_basic_logging
@@ -54,7 +54,7 @@ def _setup_cli(cli: Cli) -> None:
     _setup_llama_cli(cli)
     _setup_mt_cli(cli)
     _setup_wav2vec2_asr_cli(cli)
-    _setup_eval_cli(cli)
+    _setup_hg_cli(cli)
 
     # Set up 3rd party CLI extensions.
     for entry_point in entry_points(group="fairseq2.cli"):
